# crm/docs/openapi-mvp.yaml
openapi: 3.1.0
info:
  title: CRM Automotivo MVP API
  version: 0.1.0
  description: Contrato inicial da API Fastify do CRM automotivo MVP.
servers:
  - url: http://localhost:3333
    description: Desenvolvimento local
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Autentica usuario e cria sessao.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoga sessao atual.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Retorna usuario autenticado e permissoes efetivas.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/permissions/check:
    post:
      tags: [Auth]
      summary: Checa permissao RBAC especifica.
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "403": { $ref: "#/components/responses/Error" }
...
  /customers/{id}:
    get: { tags: [Customers], summary: Detalha cliente., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
    patch: { tags: [Customers], summary: Atualiza cliente., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
    delete: { tags: [Customers], summary: Arquiva cliente., parameters: [ { $ref: "#/components/parameters/Id" } ], responses: { "200": { $ref: "#/components/responses/Ok" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Ok:
      description: Resposta bem-sucedida.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
...
    Error:
      description: Erro padronizado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            correlationId:
              type: string